[3967] change build.scala to build.sbt #3967
